Add support for per-namespace tiller installation
Namespaces are specified as an Array, and if more than one namespace is
provided, Tiller is installed and scoped to each namespace with the
proper RBAC resources. If only one namespace is supplied, Tiller is
installed in this namespace with cluster-wide RBAC privileges.

ERB templates were converted to EPP to enable these changes.